+++ apps/student/src/main/java/com/instructure/student/features/shareextension/StudentShareExtensionActivity.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2022 - present Instructure, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3 of the License.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.instructure.student.features.shareextension
+
+import android.os.Bundle
+import com.instructure.pandautils.features.shareextension.ShareExtensionActivity
+import com.instructure.student.activity.LoginActivity
+import com.instructure.student.util.Analytics
+
+class StudentShareExtensionActivity : ShareExtensionActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        Analytics.trackAppFlow(this)
+    }
+
+    override fun exitActivity() {
+        val intent = LoginActivity.createIntent(this)
+        startActivity(intent)
+        finish()
+    }
+}
