+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/file/upload/FileUploadDialogFragment.kt
@@ -201,6 +201,7 @@ class FileUploadDialogFragment : DialogFragment() {
         intent.putExtras(bundle)
         requireActivity().startService(intent)
         dialogCallback?.invoke(EVENT_ON_UPLOAD_BEGIN)
+        dismiss()
     }
 
     private fun takePicture() {
