+++ libs/pandautils/src/main/res/layout/activity_share_file.xml
